Moved inspire_api to old_inspire_api, inspire_api now is just a switch between old and new. Minor changes to tests
